+++ src/components/eventCalendarViews/MainSearchView.tsx
@@ -1,0 +1,45 @@
+import { FormControl, FormLabel, Input, Text, VStack } from '@chakra-ui/react';
+
+import { EventCard } from './EventCard';
+import { useCalendarView } from '../../hooks/useCalendarView';
+import { useEventForm } from '../../hooks/useEventForm';
+import { useEventOperations } from '../../hooks/useEventOperations';
+import { useNotifications } from '../../hooks/useNotifications';
+import { useSearch } from '../../hooks/useSearch';
+
+export const MainSearchView = () => {
+  const { editingEvent, setEditingEvent, editEvent } = useEventForm();
+  const { events, deleteEvent } = useEventOperations(Boolean(editingEvent), () =>
+    setEditingEvent(null)
+  );
+  const { view, currentDate } = useCalendarView();
+  const { searchTerm, filteredEvents, setSearchTerm } = useSearch(events, currentDate, view);
+  const { notifiedEvents } = useNotifications(events);
+
+  return (
+    <VStack data-testid="event-list" w="500px" h="full" overflowY="auto">
+      <FormControl>
+        <FormLabel>일정 검색</FormLabel>
+        <Input
+          placeholder="검색어를 입력하세요"
+          value={searchTerm}
+          onChange={(e) => setSearchTerm(e.target.value)}
+        />
+      </FormControl>
+
+      {filteredEvents.length === 0 ? (
+        <Text>검색 결과가 없습니다.</Text>
+      ) : (
+        filteredEvents.map((event) => (
+          <EventCard
+            key={event.id}
+            event={event}
+            notifiedEvents={notifiedEvents}
+            editEvent={editEvent}
+            deleteEvent={deleteEvent}
+          />
+        ))
+      )}
+    </VStack>
+  );
+};
