+++ src/hooks/useAlert.ts
@@ -1,0 +1,7 @@
+import { useState } from 'react';
+
+export const useAlert = () => {
+  const [isOverlapDialogOpen, setIsOverlapDialogOpen] = useState(false);
+  const [overlappingEvents, setOverlappingEvents] = useState<Event[]>([]);
+  return { isOverlapDialogOpen, setIsOverlapDialogOpen, overlappingEvents, setOverlappingEvents };
+};
